Rename RANDOM_WALK_STEPS to MAX_RANDOM_WALK_STEPS
The global variable RANDOM_WALK_STEPS has been renamed to MAX_RANDOM_WALK_STEPS for clarity. This change was to better represent the purpose of the variable which is to define the maximum number of random walk steps. The reference to this variable in the condition for blue LED and incrementation of random walk steps is also updated.